Updated authentication ways
Old auth ways would use the partial content & no content status codes. I also moved probably about 90% of the auth into the Utils auth functions. Moved Codes interface from Path to Status_Codes. Fixed random quote if your on linux and running as root. Ignored yarn-error file

+++ src/Paths/API/Delete_Short.ts
@@ -2,7 +2,6 @@ import Path from '../../Structures/Path';
 import Evolve from '../../Structures/Evolve';
 import Base from '../../Structures/Base';
 import { Request, Response } from 'express';
-import { isArray } from 'util';
 
 class DeleteShort extends Path {
     constructor(evolve: Evolve, base: Base) {
@@ -14,20 +13,15 @@ class DeleteShort extends Path {
     }
 
     async execute(req: Request, res: Response): Promise<Response> {
-        if (!req.headers.token && !req.headers.uid) {
-            return res.status(this.codes.noContent).send('[ERROR] Missing authorization token and user ID!');
-        } if (!req.headers.token || !req.headers.uid) {
-            return res.status(this.codes.partialContent).send('[ERROR] Missing either authorization token or user ID!');
+        // Check auth
+        const auth = await this.Utils.authToken(req);
+        if (!auth || typeof auth === 'string') {
+            return res.status(this.codes.unauth).send(auth || '[ERROR] Authorization failed. Who are you?');
         }
-        if (isArray(req.headers.token) || isArray(req.headers.uid) ) {
-            return res.status(this.codes.badReq).send('[ERROR] Neither header auth field may be an array!');
-        }
-        const auth = await this.Utils.authToken(req.headers.token, req.headers.uid);
-        if (!auth) {
-            return res.status(this.codes.unauth).send('[ERROR] Authorization failed. Who are you?');
-        }
+
+        // Check query
         if (!req.query || !req.query.id) {
-            return res.status(this.codes.badReq).send('[ERROR] ID needed in query!');
+            return res.status(this.codes.badReq).send('[ERROR] MISSING ID!');
         }
 
         const short = await this.base.schemas.Shorten.findOneAndRemove( { ID: req.query.id } );
